requirements.txt
fastapi
uvicorn
openai
faiss-cpu
PyPDF2
python-multipart